cookiePing
this frontend commit aims to accompany the backend commit to create a cookiePing in order to prevent constant calling of refresh. It also improves logic for login and logout in order to add & clear user data and cookies

known bug: logging out from profile page should reload the page, but it messes with how cookies are managed. also, the profile page should logically have a watcher for users